refactor(bot): extract book name validation to separate module
Move `validate_and_normalize_book_name` and its associated helper logic from `bot.py` to a new `validation.py` module to improve separation of concerns and resolve circular import dependencies in `triggers.py`.

Key changes:
- Create `src/biblebot/validation.py` to house book name normalization logic.
- Remove normalization functions and `_ALL_NAMES_TO_CANONICAL` lookup from `src/biblebot/bot.py`.
- Update imports in `src/biblebot/triggers.py` to use the new validation module, removing the local import within `_match_reference`.
- Update test imports in `tests/test_bot.py` and `tests/test_matrix_utils.py` to reflect the new module structure.
